requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
airportsdata==20250706
annotated-types==0.7.0
anyio==4.10.0
astor==0.8.1
attrs==25.3.0
blake3==1.0.5
blinker==1.9.0
blockbuster==1.5.25
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
compressed-tensors==0.10.1
cryptography==44.0.3
cupy-cuda12x==13.5.1
cython==3.1.2
dacite==1.9.2
dataclasses-json==0.6.7
depyf==0.18.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email-validator==2.2.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
filelock==3.18.0
filetype==1.2.0
flask==3.1.1
flatbuffers==25.2.10
forbiddenfruit==0.1.4
frozenlist==1.7.0
fschat==0.2.36
fsspec==2025.7.0
ftfy==6.3.1
gguf==0.17.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.7.0
interegular==0.3.3
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-rs==0.29.1
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.1.9
langchain-openai==0.3.28
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.9
langgraph==0.6.3
langgraph-api==0.2.120
langgraph-checkpoint==2.1.1
langgraph-cli==0.3.6
langgraph-prebuilt==0.6.3
langgraph-runtime-inmem==0.6.8
langgraph-sdk==0.2.0
langsmith==0.4.10
lark==1.2.2
latex2mathml==3.78.0
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.11
markdown-it-py==3.0.0
markdown2==2.5.4
markupsafe==3.0.2
marshmallow==3.26.1
mcp==1.12.3
mdurl==0.1.2
mistral-common==1.8.3
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nh3==0.3.0
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.22.1
openai==1.98.0
opencv-python-headless==4.12.0.88
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-semantic-conventions-ai==0.4.12
orjson==3.11.1
ormsgpack==1.10.0
outlines==0.1.11
outlines-core==0.1.26
packaging==25.0
pandas==2.3.1
partial-json-parser==0.2.1.1.post6
peft==0.17.0
pillow==11.3.0
portalocker==3.2.0
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
prompt-toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pygments==2.19.2
pyjnius==1.6.1
pyjwt==2.10.1
pyserini==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.1
qdrant-client==1.15.1
rank-llm==0.21.0
ray==2.48.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentencepiece==0.2.0
sentry-sdk==2.34.1
setuptools==78.1.1
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.42
sse-starlette==2.1.3
starlette==0.47.2
streamlit
structlog==25.4.0
svgwrite==1.4.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.4
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.54.1
triton==3.3.0
truststore==0.10.3
typer==0.16.0
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vllm==0.9.1
watchfiles==1.1.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==15.0.1
werkzeug==3.1.3
wheel==0.45.1
xformers==0.0.30
xgrammar==0.1.19
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
